finished all tasks, checked every methods of API

+++ model/contacts.js
@@ -1,58 +1,34 @@
-// const db = require('./mongoDB/db')
-// const {ObjectID} = require('mongodb')
 const contacts = require('../model/schemas/contact')
 
-
-// const getCollection = async (db, name) => {
-//   const client = await db
-//   const collection = await client.db().collection(name)
-//   return collection
-// }
-
-const listContacts = async () => {
-  // const collection = await getCollection(db, 'contacts')
-  // const results = await collection.find().toArray()
-  const results = await contacts.find()
+const listContacts = async (userId, query) => {
+  const results = await contacts.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'subscription email',
+  })
   return results
 
 }
 
-const getContactById = async (id) => {
-  // const collection = await getCollection(db, 'contacts')
-  // const objectid = new ObjectID(id)
-  // const [result] = await collection.find({_id: objectid}).toArray()
-  const result = await contacts.findOne({_id: id})
+const getContactById = async (userId, id) => {
+  const result = await contacts.findOne({ _id: id, owner: userId })
   return result
 }
 
-const removeContact = async (id) => {
-  // const collection = await getCollection(db, 'contacts')
-  // const objectid = new ObjectID(id)
-  // const { value: result } = await collection.findOneAndDelete({ _id: objectid })
-  const result = await contacts.findByIdAndRemove({ _id: id })
+const removeContact = async (userId, id) => {
+  const result = await contacts.findByIdAndRemove({ _id: id, owner: userId })
   return result
-  
+   
 }
 
 
 
-const addContact = async (body) => {
-  // const collection = await getCollection(db, 'contacts')
-  // const recordContact = {
-  //   ...body
-  // }
-  // const  {ops: [result]}  = await collection.insertOne(recordContact)
-
-  const result = await contacts.create(body)
+const addContact = async (userId, body) => {
+  const result = await contacts.create({ ...body, owner: userId})
     return result
 }
 
-const updateContact = async (id, body) => {
-  // const collection = await getCollection(db, 'contacts')
-  // const objectid = new ObjectID(id)
-  // const { value: result } = await collection.findOneAndUpdate({ _id: objectid }, { $set: body }, { returnOriginal: false })
-
-  const result = await contacts.findByIdAndUpdate({ _id: id }, { ...body}, {new: true})
+const updateContact = async (userId, id, body) => {
+  const result = await contacts.findByIdAndUpdate({ _id: id, owner: userId }, { ...body}, {new: true})
   return result
 }
 
@@ -62,4 +38,4 @@ const updateContact = async (id, body) => {
   removeContact,
   addContact,
   updateContact,
-}
+}
